refactoring HttpClientConfiguration #21
