+++ kafka-connect-http-connectors/src/main/java/io/github/clescot/kafka/connect/sse/client/okhttp/SseConfiguration.java
@@ -4,7 +4,6 @@ import com.google.common.base.Preconditions;
 import com.launchdarkly.eventsource.*;
 import com.launchdarkly.eventsource.background.BackgroundEventSource;
 import io.github.clescot.kafka.connect.Configuration;
-import io.github.clescot.kafka.connect.http.client.HttpClientConfiguration;
 import io.github.clescot.kafka.connect.http.client.HttpClientFactory;
 import io.github.clescot.kafka.connect.http.client.okhttp.OkHttpClient;
 import io.github.clescot.kafka.connect.http.core.HttpRequest;
